+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed topic retry policy callback call: not call it with nil error 
+
 ## v3.42.10
 * Added exit from retryer if got grpc-error `Unauthenticated` on `discovery/ListEndpoints` call